requirements.txt
requests==2.32.3
beautifulsoup4==4.14.2
pdfplumber==0.11.7
anthropic==0.84.0
python-dotenv==1.1.0
pytest==9.0.2
